+++ src/lib/draw/point/PointControls.svelte
