+++ src/lib/draw/polygon/PolygonControls.svelte
