+++ src/lib/draw/polygon/polygon_tool.ts
@@ -18,8 +18,8 @@ import {
   isPoint,
   isPolygon,
   isLine,
-} from "../../maplibre_helpers";
-import { colors, circleRadius } from "../../colors";
+} from "../../../maplibre_helpers";
+import { colors, circleRadius } from "../../../colors";
 
 const source = "edit-polygon-mode";
